a11ypal_backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.3.0
python-decouple==3.8
